# Reduce tracing overhead