--- a/changelog/@unreleased/pr-1413.v2.yml
+++ b/changelog/@unreleased/pr-1413.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Remove unnecessary debug logging
+  links:
+  - https://github.com/palantir/tracing-java/pull/1413