## Before this PR

For services with very high concurrency and request volumes (e.g. timelock), there is unnecessary overhead and contention in `Tracer::logClearingTrace()` determining whether debug logging is enabled. This `Tracer::clearCurrentTrace()` debug logging has outlived its useful life.

Builds on https://github.com/palantir/tracing-java/pull/1412

```
Benchmark                             (observability)  Mode  Cnt      Score      Error  Units
TracingBenchmark.nestedSpans            DO_NOT_SAMPLE  avgt   10    494.292 ±   43.218  ns/op
TracingBenchmark.nestedSpans                UNDECIDED  avgt   10    748.402 ±   36.206  ns/op
TracingBenchmark.nestedSpans                   SAMPLE  avgt   10  30336.603 ± 5844.814  ns/op
TracingBenchmark.traceWithSingleSpan    DO_NOT_SAMPLE  avgt   10    121.955 ±   18.840  ns/op
TracingBenchmark.traceWithSingleSpan        UNDECIDED  avgt   10    124.761 ±   18.362  ns/op
TracingBenchmark.traceWithSingleSpan           SAMPLE  avgt   10    368.724 ±   76.932  ns/op
```

## After this PR
```
Benchmark                             (observability)  Mode  Cnt      Score      Error  Units
TracingBenchmark.nestedSpans            DO_NOT_SAMPLE  avgt   10    486.959 ±    3.825  ns/op
TracingBenchmark.nestedSpans                UNDECIDED  avgt   10    741.624 ±   50.668  ns/op
TracingBenchmark.nestedSpans                   SAMPLE  avgt   10  30965.189 ± 1968.021  ns/op
TracingBenchmark.traceWithSingleSpan    DO_NOT_SAMPLE  avgt   10     54.028 ±    4.108  ns/op
TracingBenchmark.traceWithSingleSpan        UNDECIDED  avgt   10    104.858 ±    3.227  ns/op
TracingBenchmark.traceWithSingleSpan           SAMPLE  avgt   10    395.974 ±    9.767  ns/op
```

==COMMIT_MSG==
Remove unnecessary debug logging
==COMMIT_MSG==

## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->